SCC: extract SccContext to specific file and refactor SccLine processing function
